requirements.txt
firebase-admin==6.1.0
google-generativeai==0.3.1
streamlit
plotly
# pandas==2.1.3
# APScheduler==3.10.4
# email-validator==2.1.0
python-dotenv
streamlit-autorefresh
# retrying==1.3.4
flask
